Update readme to include common error scenario
I'm getting a lot of questions regarding the netlify-lambda tool. While its actually less related to Netlify Lambda, but more to webpack and serverless in general, it does kill the experience for new people that are trying Netlify Functions when they face errors during the build stage.

In a default setup with functions its highly likely that people will use SDK's like Mongoose, Firebase, etc. Though its stated in this readme how to fix it for Firebase. Its not clear that adding `webpack-node-externals` fixes other issues as-well and its also not located on the `netlify-lambda build` section. People follow the docs step by step and so they get stuck when they face errors on this build stage. Solving this error is not clear for people that have no experience with Webpack and serverless hence this addition